fix: use Long type for shouldBeGreaterThan assertion in FileTransferManagerTest
The import uses kotest Long-specific extension but was called on Int.

+++ android/app/src/test/java/com/kidsync/app/sync/filetransfer/FileTransferManagerTest.kt
@@ -89,7 +89,7 @@ class FileTransferManagerTest : FunSpec({
 
         // Verify ZIP contents
         val zipBytes = outputStream.toByteArray()
-        zipBytes.size shouldBeGreaterThan 0
+        zipBytes.size.toLong() shouldBeGreaterThan 0L
 
         val zipIn = ZipInputStream(ByteArrayInputStream(zipBytes))
         val entries = mutableMapOf<String, String>()
